fix(rbac): sincronizar los permisos nuevos en cada arranque
Un permiso agregado al catálogo nunca llegaba a un entorno ya desplegado. Las
migraciones traen el esquema, pero las filas de Permission son datos y ninguna
migración las incluye; el único que las crea es `seed-rbac`, que no corre en
ningún deploy.

El síntoma de esta semana: `inspecciones:edit-finalizada` no existe en la base
de producción. La edición de inspecciones finalizadas está desplegada desde el
19/08 y es invisible incluso para un administrador — `can()` devuelve false
para un permiso inexistente, así que el botón de habilitar edición no se
renderiza nunca. El tablero de roles tampoco puede otorgarlo, porque lista los
permisos que existen en la base.

`seed-rbac` no servía para arreglarlo: reconcilia, y contra producción
revocaría las 13 asignaciones otorgadas a mano (gerente tiene 21 permisos
contra los 17 que declara el catálogo; técnico, 19 contra 10).

- El catálogo pasa a `prisma/rbac-catalog.ts`, compartido por el seed y el
  nuevo sync. `seed-rbac` mantiene su lógica y queda advertido de que revoca.
- `sync-permissions` trata el catálogo como un piso, no como un espejo: crea
  los permisos que falten, completa los roles declarados y no borra nada. Lo
  otorgado a mano sobrevive.
- Corre desde el CMD del contenedor, no desde `postinstall`: el stage `deps`
  usa `npm ci --ignore-scripts`, así que `postinstall` nunca se ejecuta en la
  imagen. También queda encadenado en `postinstall` para desarrollo local.
