+++ package.json
@@ -4,7 +4,9 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "postinstall": "prisma generate && prisma migrate deploy",
+    "postinstall": "prisma generate && prisma migrate deploy && npm run db:sync-permissions",
+    "db:sync-permissions": "tsx prisma/sync-permissions.ts",
+    "db:seed-rbac": "tsx prisma/seed-rbac.ts",
     "build": "next build",
     "start": "next start",
     "lint": "eslint",
